refactor: rename ModelUtil to AuxUtil
